read_and_process/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-qdrant
langchain-huggingface
qdrant-client
sentence-transformers
transformers
Pillow
PyMuPDF
python-dotenv
torch
fastembed
python-multipart